Move TextMultilineModifiable to ImGuiXtra
